Expand library interface
This project was mostly a meme but now keyborb wants to actually use it
for something.

The library now provides the weather info as a table which additionally
contains a field for the intensity. You can poll the current weather at
any point via the new GetCurrentWeatherInfo function. Additionally,
we'll try and persist knowledge of the weather across UI reloads.